+++ tools/CMakeLists.txt
@@ -5,7 +5,7 @@ include(../cmake/HermitCore-Paths.cmake)
 
 add_compile_options(-std=c99)
 
-add_executable(proxy proxy.c utils.c uhyve.c uhyve-net.c)
+add_executable(proxy proxy.c utils.c uhyve.c uhyve-net.c uhyve-x86_64.c uhyve-aarch64.c)
 target_compile_options(proxy PUBLIC -pthread)
 target_compile_options(proxy PUBLIC -DMAX_ARGC_ENVC=${MAX_ARGC_ENVC})
 target_link_libraries(proxy -pthread)
